+++ src/handlers/equipments/deleteEquipment.ts
@@ -1,6 +1,6 @@
 import type { RouteHandler } from "@hono/zod-openapi";
 import type { Context } from "hono";
-import type { deleteEquipmentsRoute } from "~/routers/equipments/equipments";
+import type { deleteEquipmentsRoute } from "~/routers/equipments/route";
 
 export const deleteEquipmentHandler: RouteHandler<
 	typeof deleteEquipmentsRoute
