+++ src/handlers/users/createUser.ts
@@ -2,7 +2,7 @@ import type { RouteHandler } from "@hono/zod-openapi";
 
 import type { Context } from "hono";
 import { insertUser } from "~/models/user";
-import type { createUserRoute } from "~/routers/users/users";
+import type { createUserRoute } from "~/routers/users/route";
 import { createUserRequestSchema } from "~/schema/user";
 
 import { validateRequestBody } from "~/utils/validateRequestBody";
