+++ src/handlers/users/rejectUser.ts
@@ -1,7 +1,7 @@
 import type { RouteHandler } from "@hono/zod-openapi";
 import type { Context } from "hono";
 import { updateUserById } from "~/models/user";
-import type { rejectUserRoute } from "~/routers/users/users";
+import type { rejectUserRoute } from "~/routers/users/route";
 import { rejectUserPathParamSchema } from "~/schema/user";
 
 export const rejectUserHandler: RouteHandler<typeof rejectUserRoute> = async (
