+++ src/routers/equipments/route.ts
